moved code and fixed problem when there are no css files
