+++ packages/snap/src/generate-snapshots/generate-snapshots.ts
@@ -9,8 +9,6 @@ import {consoleLog, consoleWarn} from '@ui-autotools/utils';
 import {dedent} from './dedent';
 import {parseSnapshotFilename} from './filename-utils';
 import { IFileInfo } from './build-base-files';
-import {generateSnapshotFilename} from './filename-utils';
-import {writeDataToFs} from './write-data-to-fs';
 
 function findComponentByName(name: string, Registry: IRegistry): IComponentMetadata<any, any> | void {
     // We only have to do this because we currently map the component definitions to their metadata,
@@ -145,48 +143,3 @@ export interface IFileInfo {
   filepath: string;
   data: string;
 }
-
-export const generateSnapshots2 = async (projectDir: string, tempDirectory: string, Registry: IRegistry) => {
-  const files: IFileInfo[] = [];
-  consoleLog('Generating snapshots...');
-  Registry.metadata.components.forEach((componentMetadata) => {
-    const simIndex = componentMetadata.simulations.length;
-    const compName = componentMetadata.exportName;
-
-    if (compName) {
-      for (let i = 0; i < simIndex; i++) {
-        const simulationName = componentMetadata.simulations[i].title;
-        const basename = generateSnapshotFilename(compName, simulationName, i);
-        const filepath = path.join(tempDirectory, basename + '.snapshot.ts');
-        const data = createHtml(projectDir, componentMetadata, componentMetadata.simulations[i].props);
-        files.push({basename, filepath, data});
-      }
-    }
-  });
-
-  writeDataToFs(files);
-  return files;
-};
-
-const createHtml = (projectDir: string, compMetadata: IComponentMetadata<any, any>, props: any) => {
-  if (!compMetadata.compPath) {
-    throw new Error(`Component ${compMetadata.exportName} has no path`);
-  }
-  const comp = require(path.join(projectDir, compMetadata.compPath)).default;
-  const cssLink = compMetadata.cssPath ? `<link rel="stylesheet" type="text/css" href="${path.join(projectDir, compMetadata.cssPath)}.css">` : '';
-  const componentString = renderToStaticMarkup(React.createElement(comp, props));
-
-  return `<!DOCTYPE html>
-  <html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>${compMetadata.exportName}</title>
-    ${cssLink}
-  </head>
-  <body>
-    ${componentString}
-  </body>
-  </html>`;
-};
